+++ modules/Accounting/Incomes.php
@@ -23,6 +23,19 @@ if ( ! empty( $_REQUEST['values'] )
 		{
 			$sql = "UPDATE accounting_incomes SET ";
 
+			$columns['FILE_ATTACHED'] = _saveIncomesFile( $id );
+
+			if ( ! $columns['FILE_ATTACHED'] )
+			{
+				unset( $columns['FILE_ATTACHED'] );
+
+				if ( empty( $columns ) )
+				{
+					// No file, and FILE_ATTACHED was the only column, skip.
+					continue;
+				}
+			}
+
 			foreach ( (array) $columns as $column => $value )
 			{
 				$sql .= DBEscapeIdentifier( $column ) . "='" . $value . "',";
@@ -39,19 +52,7 @@ if ( ! empty( $_REQUEST['values'] )
 			$fields = 'SCHOOL_ID,SYEAR,';
 			$values = "'" . UserSchool() . "','" . UserSyear() . "',";
 
-			if ( isset( $_FILES['FILE_ATTACHED'] ) )
-			{
-				$columns['FILE_ATTACHED'] = FileUpload(
-					'FILE_ATTACHED',
-					$FileUploadsPath . UserSyear() . '/staff_' . User( 'STAFF_ID' ) . '/',
-					FileExtensionWhiteList(),
-					0,
-					$error
-				);
-
-				// Fix SQL error when quote in uploaded file name.
-				$columns['FILE_ATTACHED'] = DBEscapeString( $columns['FILE_ATTACHED'] );
-			}
+			$columns['FILE_ATTACHED'] = _saveIncomesFile( $id );
 
 			$go = 0;