+++ CHANGES.md
@@ -13,6 +13,7 @@ Changes in 10.4
 - Add optional $id param to FilesUploadUpdate() in FileUpload.fnc.php
 - JS Only show laoding spinner if file input has selected files in warehouse.js
 - Add File Attached Input for existing Fees/Payments in StudentFees.php, StudentPayments.php & Student_Billing/functions.inc.php
+- Add File Attached Input for existing Salaries/Staff Payments/Incomes/Expenses in Expenses.php, Incomes.php, Salaries, StaffPayments.php & Accounting/functions.inc.php
 
 Changes in 10.3.2
 -----------------